Move code from `api.rs` to `util.rs`
